Updated Querying and Search Activities
Sorry for lack of comments, focusing on Functionality.

Querying for everything we want is set up as you will see in search activity. However, some logic is broken. The Search activity list view only works because the edit text user_search is prefilled.

The desirable behavior is that the loader won't launch until the user_search is filled.