I made some cosmetic changes!
Changed activity_form_visited to activity_form since we only use this one layout for both forms.

deleted activity_form_want_to_visit because we are not using it

added a string for the search button

removed the buttons in the form layout and put new icons on the menu buttons for adding/updating and deleting the form item

updated the search list view. I added some new strings and colors for this purpose. I also just used the same theme from the other listView.

Changed the lsit view theme back to the previous version because I could not kill the orange that appears for no reason